Fix arguments referenced in wrong environment
A reference in the function's alist used to get referred in the closure
rather than called env.

The `alist` is constructed in such a way that all bindings refer to
concrete values, rather than other references.

The alist of the form `((x a))`, rather than `((x 42))` will cause `a`
to be looked up in the function closure, rather than the caller's
environment. This is prevented by `resolving` the value of each argument
to a value other than an atom before zipping with the formal arguments.

This gives the added benefit that the caller's environment is not needed
while evaluating the function body, preventing behavior similar to
dynamic scoping.