Handle errors gracefully with an either type, WIP
Wrapped the result of evaluation in an Either type to avoid a lot of issues,
including unhandled repl crashes, poor semantics etc.

In general the code is more robust, but lot more uglier with way too many error
checks. A transformer stack of Either and State should clean things up.